// packages/telemetry/types/src/telemetry.ts
export enum SessionState {
  invalid = 0,
  get_in_car = 1,
  warmup = 2,
  parade_laps = 3,
  racing = 4,
  checkered = 5,
  cool_down = 6
}
/**
 * This interface was referenced by `Telemetry`'s JSON-Schema
 * via the `definition` "TrackLocation".
 */
export enum TrackLocation {
  not_in_world = -1,
  off_track = 0,
  in_pit_stall = 1,
  aproaching_pits = 2,
  on_track = 3
}
// ...
export enum PitServiceStatus {
  none = 0,
  in_progress = 1,
  complete = 2,
  too_far_left = 100,
  too_far_right = 101,
  too_far_forward = 102,
  too_far_back = 103,
  bad_angle = 104,
  cant_fix_that = 105
}
/**
 * This interface was referenced by `Telemetry`'s JSON-Schema
 * via the `definition` "Flags".
 */
export enum Flags {
  checkered = 1,
  white = 2,
  green = 4,
  yellow = 8,
  red = 16,
  blue = 32,
  debris = 64,
  crossed = 128,
  yellow_waving = 256,
  one_lap_to_green = 512,
  green_held = 1024,
  ten_to_go = 2048,
  five_to_go = 4096,
  random_waving = 8192,
  caution = 16384,
  caution_waving = 32768,
  black = 65536,
  disqualify = 131072,
  servicible = 262144,
  furled = 524288,
  repair = 1048576,
  start_hidden = 268435456,
  start_ready = 536870912,
  start_set = 1073741824,
  start_go = 2147483648
}
/**
 * This interface was referenced by `Telemetry`'s JSON-Schema
 * via the `definition` "PaceMode".
 */
export enum PaceMode {
  single_file_start = 0,
  double_file_start = 1,
  single_file_restart = 2,
  double_file_restart = 3,
  not_pacing = 4
}
/**
 * This interface was referenced by `Telemetry`'s JSON-Schema
 * via the `definition` "PaceFlags".
 */
export enum PaceFlags {
  end_of_line = 1,
  free_pass = 2,
  waved_around = 4
}